+++ packages/clusterer/directives.ts
@@ -1,0 +1,1 @@
+export {AgmMarkerCluster} from './directives/cluster';
