+++ packages/clusterer/package.tpl.json
@@ -1,0 +1,29 @@
+{
+  "name": "@agm/clusterer",
+  "version": "INSERT_HERE_VIA_BUILD_PROCESS",
+  "private": true,
+  "description": "Angular Google Maps (AGM) package for Snazzy Info Window support",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/SebastianM/angular-google-maps.git"
+  },
+  "keywords": [
+    "angular-google-maps",
+    "angular",
+    "js-marker-clusterer",
+    "info-window",
+    "google-maps",
+    "agm"
+  ],
+  "author": "Jens Fahnenbruck <jigfox@me.com>",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/SebastianM/angular-google-maps/issues"
+  },
+  "peerDependencies": {
+    "@angular/core": "^4.0.0",
+    "@agm/core": "^1.0.0-beta.0",
+    "js-marker-clusterer": "^1.0.0"
+  },
+  "homepage": "https://github.com/SebastianM/angular-google-maps#readme"
+}
