+++ packages/clusterer/services.ts
@@ -1,0 +1,1 @@
+export {ClusterManager} from './services/managers/cluster-manager';
