+++ packages/core/core.module.ts
@@ -3,7 +3,6 @@ import {AgmMap} from './directives/map';
 import {AgmCircle} from './directives/circle';
 import {AgmInfoWindow} from './directives/info-window';
 import {AgmMarker} from './directives/marker';
-import {AgmMarkerCluster} from './directives/cluster';
 import {AgmPolygon} from './directives/polygon';
 import {AgmPolyline} from './directives/polyline';
 import {AgmPolylinePoint} from './directives/polyline-point';
@@ -19,7 +18,7 @@ import {BROWSER_GLOBALS_PROVIDERS} from './utils/browser-globals';
  */
 export function coreDirectives() {
   return [
-    AgmMap, AgmMarker, AgmMarkerCluster, AgmInfoWindow, AgmCircle,
+    AgmMap, AgmMarker, AgmInfoWindow, AgmCircle,
     AgmPolygon, AgmPolyline, AgmPolylinePoint, AgmKmlLayer,
     AgmDataLayer
   ];
