+++ packages/core/directives.ts
@@ -7,4 +7,3 @@ export {AgmMarker} from './directives/marker';
 export {AgmPolygon} from './directives/polygon';
 export {AgmPolyline} from './directives/polyline';
 export {AgmPolylinePoint} from './directives/polyline-point';
-export {AgmMarkerCluster} from './directives/cluster';
