+++ packages/core/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "@agm/core",
+  "version": "1"
+}
