+++ packages/core/services.ts
@@ -6,7 +6,6 @@ export {PolygonManager} from './services/managers/polygon-manager';
 export {PolylineManager} from './services/managers/polyline-manager';
 export {KmlLayerManager} from './services/managers/kml-layer-manager';
 export {DataLayerManager} from './services/managers/data-layer-manager';
-export {ClusterManager} from './services/managers/cluster-manager';
 export {GoogleMapsScriptProtocol, LAZY_MAPS_API_CONFIG, LazyMapsAPILoader, LazyMapsAPILoaderConfigLiteral} from './services/maps-api-loader/lazy-maps-api-loader';
 export {MapsAPILoader} from './services/maps-api-loader/maps-api-loader';
 export {NoOpMapsAPILoader} from './services/maps-api-loader/noop-maps-api-loader';
