+++ scripts/packages.js
@@ -1,7 +1,8 @@
 // NPM packages (without the org name) that we publish
 const packages = [
   'core',
-  'snazzy-info-window'
+  'snazzy-info-window',
+  'clusterer',
 ];
 
 module.exports = packages;
